+++ flaskr/resources/file_transfer.py
@@ -1,4 +1,5 @@
 from flask_restful import Resource
+
 from flaskr.services.file import FileModelService
 
 class FileTransferResource(Resource):
